+++ app/javascript/vue/tasks/images/new_image/components/sqed/repository.vue
@@ -1,16 +1,17 @@
 <template>
   <fieldset class="fieldset">
     <legend>Repository</legend>
-    <div class="horizontal-left-content align-start separate-bottom">
+    <div class="horizontal-left-content align-start separate-bottom align-start">
       <smart-selector
-        class="full_width"
+        class="full_width margin-small-right"
         ref="smartSelector"
         model="repositories"
         target="CollectionObject"
         klass="CollectionObject"
         pin-section="Repositories"
         pin-type="Repository"
-        @selected="setRepository"/>
+        v-model="repository"/>
+      <lock-component v-model="settings.lock.repository"/>
     </div>
     <template v-if="repository">
       <div class="middle separate-top">
@@ -18,7 +19,7 @@
         <span class="separate-right"> {{ repository.name }}</span>
         <span
           class="circle-button button-default btn-undo"
-          @click="unsetRepository()"/>
+          @click="setRepository()"/>
       </div>
     </template>
   </fieldset>
@@ -29,9 +30,13 @@
 import { MutationNames } from '../../store/mutations/mutations'
 import { GetterNames } from '../../store/getters/getters'
 import SmartSelector from 'components/smartSelector'
+import LockComponent from 'components/lock'
 
 export default {
-  components: { SmartSelector },
+  components: {
+    SmartSelector,
+    LockComponent
+  },
 
   computed: {
     repository: {
@@ -40,6 +45,15 @@ export default {
       },
       set (value) {
         this.$store.commit(MutationNames.SetRepository, value)
+      }
+    },
+
+    settings: {
+      get () {
+        return this.$store.getters[GetterNames.GetSettings]
+      },
+      set (value) {
+        this.$store.commit(MutationNames.SetSettings, value)
       }
     }
   },