+++ app/javascript/vue/tasks/images/new_image/store/mutations/resetStore.js
@@ -1,5 +1,11 @@
 import { makeInitialState } from '../store'
 
 export default (state) => {
-  Object.assign(state, makeInitialState())
+  const newState = makeInitialState()
+  newState.settings = state.settings
+  const locked = Object.keys(state.settings.lock).filter(key => state.settings.lock[key])
+  locked.forEach(key => {
+    newState[key] = state[key]
+  })
+  Object.assign(state, newState)
 }