+++ app/javascript/vue/tasks/images/new_image/const/makeTaxonDetermination.js
@@ -1,0 +1,8 @@
+export default () => ({
+  biological_collection_object_id: undefined,
+  roles_attributes: [],
+  otu_id: undefined,
+  day_made: undefined,
+  month_made: undefined,
+  year_made: undefined
+})