+++ app/javascript/vue/tasks/sources/new_source/store/actions/removeDocumentation.js
@@ -1,0 +1,8 @@
+import { RemoveDocumentation } from '../../request/resources'
+export default ({ state }, documentation) => {
+  RemoveDocumentation(documentation.id).then(() => {
+    const index = state.documentations.findIndex(item => item.id === documentation.id)
+
+    state.documentations.splice(index, 1)
+  })
+}