+++ CHANGELOG.md
@@ -9,12 +9,13 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 - Determination, OTU and repository smart selectors on New image task [#2101]
+- Georeferences coordinates in label generate on New collecting event [#2107]
 
 ### Fixed
-
 - Not being able to get pinboard items on some circumstances
 
 [#2101]: https://github.com/SpeciesFileGroup/taxonworks/issues/2101
+[#2107]: https://github.com/SpeciesFileGroup/taxonworks/issues/2107
 
 ## [0.16.4] - 2020-03-09