+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 - Determination, OTU and repository smart selectors on New image task [#2101]
 - Georeferences coordinates in label generate on New collecting event [#2107]
+- Lock buttons on New image [#2101]
 
 ### Changed
 - Lock background color [#2112]